+++ interceptors/session.js
@@ -12,9 +12,7 @@
 "use strict";
 
 const {resolve} = require("path");
-const tools = resolve(__dirname, "..", "tools");
-const [{decipher, cipher}, key, redisService, {randomCode}, errors] = [require(resolve(tools, "security")), "cynomys-portal", require(resolve(__dirname, "..", "services", "redisService")),
-  require(resolve(tools, "utilities")), require(resolve(__dirname, "..", "errors"))];
+const [key, {removeSession, getSession, upDateSession, setSession}] = ["cynomys-portal", require(resolve(__dirname, "..", "services", "sessionService"))];
 
 const parseLanguage = header => {
   let {language = header["accept-language"]} = header;
@@ -33,14 +31,15 @@ const parseLanguage = header => {
 
 module.exports = function* (next) {
   let [start, cookie, requestIp = "0.0.0.0", header] = [Date.now(), this.cookies.get(key), this.request.ip, this.request.header];
-  let {token} = header;
-  this.requestIp = requestIp; // 绑定请求Ip到 this对象上
-  this.language = parseLanguage(header); // 绑定请求语言
+  const {token} = header;
   if (!cookie) {
     cookie = token;
   }
 
   this.token = cookie;
+  this.requestIp = requestIp; // 绑定请求Ip到 this对象上
+  this.language = parseLanguage(header); // 绑定请求语言
+
   this.success = ctx => ({
     code: 200,
     error: null,
@@ -54,46 +53,21 @@ module.exports = function* (next) {
   });
 
   this.cancel = function* () {
-    let [, equipmentId, userId] = security.decipher(cookie).split(":"); // 解构 设备id,用户id,ip
-    if (equipmentId !== this.equipmentId) {
-      errors.throwSessionError(this.language);
-    }
-    yield redisService.del(`${equipmentId}:${userId}`);
-    this.cookies.set(key, requestIp);
+    return yield removeSession(cookie);
   };
 
-  this.getSession = function* (token) {
-    try {
-			if(token) {
-        cookie = token;
-      }
-      const [permit, to] = security.decipher(cookie).split(":"); // 解构 设备id,用户id,ip
-      let session = yield redisService.get(`${to}-${permit}`);
-      return session;
-    } catch (error) {
-      return null;
-    }
+  this.getSession = function* () {
+    return yield getSession(cookie);
   };
 
   this.upDateSession = function* (value) {
-    const [permit, to] = security.decipher(cookie).split(":"); // 解构 设备id,用户id,ip
-    const _ = yield redisService.get(`${to}-${permit}`);
-    Object.assign(session, value);
-    return yield redisService.set(`${to}-${permit}`, session);
+    return yield upDateSession(cookie, value);
   };
 
-  this.setSession = function* (seems, token = cookie) {
-    const [permit, to] = security.decipher(token).split(":");
-    Object.assign(seems, {
-      secret: permit,
-      timestamp: Date.now(),
-    });
-    return yield redisService.set(`${to}-${permit}`, seems);
-  };
-
-  this.sign = (to, permit) => {
-    cookie = security.cipher(`${permit}:${to}`);
+  this.sign = function* (to, permit) {
+    cookie = sign(to, permit);
     this.cookies.set(key, cookie);
+    yield setSession(cookie, {});
     return cookie;
   };